+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+- Add support for Django and Flask Splinter browsers, that don't have a driver
+  attribute `#146 <https://github.com/pytest-dev/pytest-splinter/issues/146>`_
+  (michelts)
+
 2.0.1
 -----